Separate dependencies and lightweight test env for Python
